Keep looking if we haven't loaded the correct file
Checking the existence of the class by calling class_exists() (with
autoloading set to false, to avoid falling into the same pit) ensures
we don't bail out before having loaded the correct class.

Previously, if two paths in the include_paths contained the same file,
we would only load the first one. Since the class we were looking for
didn't exist, and we're calling class_parents() on it with the
autoload flag enabled, we ended up in the same situation again.